+++ src/contracts/SnappyImage.php
@@ -10,6 +10,7 @@
 
 namespace enupal\snapshot\contracts;
 
+use craft\web\Response;
 use Knp\Snappy\GeneratorInterface;
 use Knp\Snappy\Image;
 use enupal\snapshot\Snapshot;
@@ -25,8 +26,7 @@ class SnappyImage extends BaseSnappy
      */
     protected function getBinary()
     {
-        $plugin = Craft::$app->getPlugins()->getPlugin('enupal-snapshot');
-        $this->pluginSettings = $plugin->getSettings();
+        $this->pluginSettings = Snapshot::$app->settings->getSettings();
 
         $this->binary = '"'.$this->pluginSettings->imageBinPath.'"';
 
@@ -47,11 +47,10 @@ class SnappyImage extends BaseSnappy
 
     /**
      * @param string $html
-     * @param array  $settings display inline | url
-     *                         *
+     * @param array $settings display inline | url
      *
      * @return Response|string
-     * @throws \Exception
+     * @throws \Throwable
      * @throws \yii\web\ServerErrorHttpException
      */
     public function displayHtml($html, $settings = null)
@@ -65,11 +64,12 @@ class SnappyImage extends BaseSnappy
 
     /**
      * @param string $template
-     * @param array  $settings display inline | url
+     * @param array $settings display inline | url
      *
      * @return string
-     * @throws \Twig_Error_Loader
+     * @throws \Throwable
      * @throws \yii\base\Exception
+     * @throws \yii\web\ServerErrorHttpException
      */
     public function displayTemplate($template, $settings = null)
     {
@@ -86,10 +86,10 @@ class SnappyImage extends BaseSnappy
 
     /**
      * @param string $url
-     * @param array  $settings display inline | url
+     * @param array $settings display inline | url
      *
      * @return string
-     * @throws \Exception
+     * @throws \Throwable
      * @throws \yii\web\ServerErrorHttpException
      */
     public function displayUrl($url, $settings = null)
@@ -115,12 +115,13 @@ class SnappyImage extends BaseSnappy
     /**
      * Generate image from html or urls
      *
-     * @param string         $source Html or Urls
+     * @param string $source Html or Urls
      * @param SnappySettings $settingsModel
      *
-     * @param bool           $sourceIsHtml
+     * @param bool $sourceIsHtml
      *
      * @return string|Response
+     * @throws \Throwable
      */
     private function _generateImage($source, SnappySettings $settingsModel, $sourceIsHtml = true)
     {
@@ -136,11 +137,14 @@ class SnappyImage extends BaseSnappy
                 Snapshot::error(Snapshot::t("Unable to find the Image file: ".$settingsModel->path));
                 return Snapshot::t("Unable to display Image file on browser");
             }
-        } catch (\RuntimeException $e) {
+
+            $asset = $this->getAsset($settingsModel);
+
+        } catch (\Exception $e) {
             Snapshot::error(Snapshot::t("Something went wrong when creating the Image file: ".$e->getMessage()));
             return Snapshot::t("Something went wrong when creating the Image file, please check your logs");
         }
-        // return download link always for images
-        return $this->getPublicUrl($settingsModel->filename);
+
+        return $asset->getUrl();
     }
 }