+++ src/models/Settings.php
@@ -39,8 +39,25 @@ class Settings extends Model
      */
     public $timeout;
 
-    // Public Methods
-    // =========================================================================
+    /**
+     * @var int
+     */
+    public $volumeId;
+
+    /**
+     * @var string
+     */
+    public $singleUploadLocationSource;
+
+    /**
+     * @var string
+     */
+    public $singleUploadLocationSubpath;
+
+    /**
+     * @var bool
+     */
+    public $overrideFile = true;
 
     /**
      * @inheritdoc